requirements.txt
flask
werkzeug
web3
mysql-connector-python
qrcode
pillow
python-dotenv

# pip install -r requirements.txt
